+++ presentation/src/main/java/com/knocklock/presentation/home/menu/HomeMenu.kt
@@ -7,5 +7,5 @@ enum class HomeMenu(val icon: ImageVector) {
     CLEAR(icon = KnockLockIcons.delete),
     SAVE(icon = KnockLockIcons.save),
     SETTING(icon = KnockLockIcons.settings),
-    TMP(icon = KnockLockIcons.addBox)
+    EDIT(icon = KnockLockIcons.addBox)
 }
