+++ presentation/src/main/java/com/knocklock/presentation/lockscreen/LockScreenActivity.kt
@@ -14,15 +14,14 @@ import android.view.Gravity
 import android.view.View
 import android.view.WindowManager
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
 import androidx.activity.setViewTreeOnBackPressedDispatcherOwner
 import androidx.activity.viewModels
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.compositionContext
 import androidx.compose.ui.platform.createLifecycleAwareWindowRecomposer
-import androidx.lifecycle.ViewTreeLifecycleOwner
-import androidx.lifecycle.ViewTreeViewModelStoreOwner
+import androidx.lifecycle.setViewTreeLifecycleOwner
+import androidx.lifecycle.setViewTreeViewModelStoreOwner
 import androidx.savedstate.setViewTreeSavedStateRegistryOwner
 import com.knocklock.presentation.lockscreen.service.LockScreenNotificationListener
 import dagger.hilt.android.AndroidEntryPoint
@@ -103,8 +102,8 @@ class LockScreenActivity : ComponentActivity() {
     @OptIn(ExperimentalComposeUiApi::class)
     private fun initViewTreeOwner(composeView: ComposeView) {
         composeView.apply {
-            ViewTreeLifecycleOwner.set(this, this@LockScreenActivity)
-            ViewTreeViewModelStoreOwner.set(this, this@LockScreenActivity)
+            setViewTreeLifecycleOwner(this@LockScreenActivity)
+            setViewTreeViewModelStoreOwner(this@LockScreenActivity)
             setViewTreeSavedStateRegistryOwner(this@LockScreenActivity)
             setViewTreeOnBackPressedDispatcherOwner(this@LockScreenActivity)
             compositionContext = createLifecycleAwareWindowRecomposer(lifecycle = lifecycle)
